rename test pages components to avoid duplicates with examples
