+++ examples/src/pages/e2e/cad/ClippingTestPage.tsx
