+++ examples/src/pages/e2e/cad/DefaultCadTestPage.tsx
