+++ examples/src/pages/e2e/cad/DefaultCameraTestPage.tsx
