+++ examples/src/pages/e2e/cad/GhostModeTestPage.tsx
