+++ examples/src/pages/e2e/cad/HighlightTestPage.tsx
