+++ examples/src/pages/e2e/cad/ScaledModelTestPage.tsx
